Fixed the delete bug

+++ MobileApp_C971_Chakradhar_Konala/MobileApp_C971_Chakradhar_Konala/EditTermPage.xaml.cs
@@ -76,13 +76,18 @@ namespace MobileApp_C971_Chakradhar_Konala
 
         private async void DeleteTerm_OnClicked(object sender, EventArgs e)
         {
-            var getCourseWithTermID = await App.Database.GetCourseAsync(term.ID);
-            foreach (var Course in getCourseWithTermID)
+            if (await DisplayAlert("Confirmation",
+                "This will delete Term, related courses and assesments. Are you sure?", "Yes", "No"))
             {
-                await App.Database.DeleteCourseAsync(Course);
+                var getCourseWithTermID = await App.Database.GetCourseAsync(term.ID);
+                foreach (var Course in getCourseWithTermID)
+                {
+                    await App.Database.DeleteCourseAsync(Course);
+                }
+
+                await App.Database.DeleteTermAsync(term);
+                await Navigation.PopToRootAsync();
             }
-            await App.Database.DeleteTermAsync(term);
-            await Navigation.PopToRootAsync();
         }
     }
 }
